requirements.txt
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
install==1.3.5
packaging==21.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
pyparsing==3.0.8
python-dotenv==0.20.0
python-slugify==6.1.1
redis==4.2.2
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.0
